Add new feature Visitor View

+++ app/helpers/images_helper.rb
@@ -1,7 +1,7 @@
 module ImagesHelper
 
   def show_delete_if_logged_in(image)
-    if current_user
+    if current_user and !visitor_view
       link_to 'X', image, :data => {confirm: 'Are you sure?'}, method: :delete, :remote => true, :class=>'delete-image white small button no-print'
     end
   end
